git-delete-merged bugfixes
Added --merge-branch option (with git defaults) to signal what the main branch is when comparing what has been merged.
Handle the case of the current branch not being even with the remote branch or the remote being gone.

Breaking:
changed the section in the git config where the props live

Added some more minor tests:

$  self_test=1 ~/bin/git-delete-merged
 #filter-branches() test
    filter test1        : OK
    filter test2        : OK
    filter test3        : OK

 #skipped() test
    verify defaults     : OK
    verify override     : OK