macOS/BSD compatible awk fixes
In GNU Awk, match(string, regex, array) will populate the array with the full match and all submatches:

array[0] = full match

array[1], array[2], etc. = capture groups

But in BSD Awk, including the version on macOS, this feature is not supported. match() simply returns the position of the match, and does not populate the third argument with submatches